requirements.txt
GeoNode==2.10
django-overextends
django-annoying
django-rosetta==0.9.3
django-grappelli==2.12.3
djproxy
simplejson
django-analytical==2.5.0
django-taggit-templatetags2
